+++ test/test_nuke.py
@@ -1,13 +1,18 @@
+# coding: utf-8
 import sys
+import unittest
 
 sys.path.append('../yurlungur')
 import yurlungur as yr
 
-reload(yr)
 
-yr.YNode("").create("TimeWarp")
+class TestResolve(unittest.TestSuite):
 
-if hasattr(meta, "knob"):
-    b = nuke.toNode("aaa")
-    k = nuke.Array_Knob("name", "label")
-    b.addKnob(k)
+    def test_node(self):
+        yr.YNode("").create("TimeWarp")
+
+        if hasattr(yr.meta, "knob"):
+            import nuke
+            b = nuke.toNode("aaa")
+            k = nuke.Array_Knob("name", "label")
+            b.addKnob(k)
