--- yurlungur/adapters/ue.py
@@ -1,53 +1,0 @@
-# coding: utf-8
-import unreal
-
-
-@unreal.uclass()
-class EditorUtil(unreal.GlobalEditorUtilityBase):
-    pass
-
-
-@unreal.uclass()
-class GetEditorAssetLibrary(unreal.EditorAssetLibrary):
-    pass
-
-
-@unreal.uclass()
-class MaterialEditingLib(unreal.MaterialEditingLibrary):
-    pass
-
-
-@unreal.uclass()
-class GetAnimationLibrary(unreal.AnimationLibrary):
-    pass
-
-
-class Project(object):
-    def __init__(self, project):
-        self.project = project
-
-    @property
-    def sequences(self):
-        return Timeline(self.project)
-
-
-class Timeline(object):
-    def __init__(self, timeline):
-        self.timeline = timeline
-
-    @property
-    def tracks(self):
-        return Track(self.timeline)
-
-
-class Track(object):
-    def __init__(self, track):
-        self.track = track
-
-    @property
-    def clips(self):
-        return Item(self.track)
-
-
-class Item(object):
-    pass
